# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: SmartCart AI - API Specification
  description: |
    Complete API documentation for SmartCart AI platform - an AI-powered supermarket ordering system.
    
    ## Architecture
    - Microservices-based architecture
    - RESTful APIs with JSON payloads
    - JWT-based authentication
    - Rate limiting enabled
    
    ## Base URLs
    - Production: `https://api.smartcart.ai`
    - Staging: `https://api-staging.smartcart.ai`
    - Development: `http://localhost:8080`
    
  version: 1.0.0
  contact:
    name: SmartCart AI Team
    email: api@smartcart.ai
  license:
    name: Proprietary
servers:
  - url: https://api.smartcart.ai/v1
    description: Production server
  - url: https://api-staging.smartcart.ai/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: List all products
      description: Get paginated list of products with optional filters
      parameters:
        - name: store_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: category_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          description: Search query for product name/description
          schema:
            type: string
        - name: in_stock
          in: query
          description: Filter by stock availability
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    
    post:
      tags:
        - Products
      summary: Create new product
      description: Add a new product to the catalog (Admin/Owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
  /products/{product_id}:
    get:
      tags:
        - Products
      summary: Get product details
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Products
      summary: Update product
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Product'
    
    delete:
      tags:
        - Products
      summary: Delete product (soft delete)
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/verify-otp
  schemas:
...
    Product:
      type: object
      properties:
        product_id:
          type: integer
        store_id:
          type: string
          format: uuid
        category_id:
          type: integer
        product_name:
          type: string
        description:
          type: string
        brand:
          type: string
        mrp:
          type: number
        selling_price:
          type: number
        discount_percent:
          type: number
        stock_quantity:
          type: number
        unit_type:
          type: string
          enum: [kg, gram, liter, ml, piece, dozen, packet]
        min_stock_level:
          type: number
        image_url:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    ProductInput:
      type: object
      required:
        - store_id
        - product_name
        - selling_price
        - unit_type
      properties:
        store_id:
          type: string
          format: uuid
        category_id:
          type: integer
        product_name:
          type: string
        description:
          type: string
        brand:
          type: string
        mrp:
          type: number
        selling_price:
          type: number
        stock_quantity:
          type: number
        unit_type:
          type: string
        min_stock_level:
          type: number
        image_url:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...